+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 <!-- `virtual-scroller`: in `.updateItems()` handle a case when `items.length` is the same, in which case find different items and if those items are rendered then maybe update them on screen and update their height, if the items are past rendered then maybe just discard all item heights past rendered, if the items are before rendered then maybe ignore and it will jump on scroll up which is kinda acceptable. -->
 
+1.6.6 / 18.12.2020
+==================
+
+* Removed `getItemKey(item)` property of React `<VirtualScroller/>`. Use `getItemId(item)` option of `VirtualScroller` instead.
+
+* Added `getItemId(item)` option of `VirtualScroller`.
+
+1.6.5 / 17.12.2020
+==================
+
+* Added `getItemKey(item)` property of React `<VirtualScroller/>`: that fixes forced re-rendering of all visible items whenever `items` property changes.
+
 1.6.4 / 12.12.2020
 ==================